Refactors game code into a separate model.
Moves the code for running a game into
the game/base module and the PyGame portion
into the game/visual module. This will
help in the eventual plan to string together
multiple games in a "X out of Y" match, and
those matches into a tournament bracket.
Additionally, this will help in creating a
version of the game that does not take the
time to create the visual output of the game.